refactor(compiler-core): move compile into separate file
